backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.12.1
boto3==1.35.0
botocore==1.35.99
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Deprecated==1.3.1
ecdsa==0.19.1
fastapi==0.115.0
fastapi-csrf-protect==1.0.7
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.0
idna==3.11
itsdangerous==2.2.0
jmespath==1.1.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.10
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.6.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-magic-bin==0.4.14
python-multipart==0.0.12
PyYAML==6.0.3
rsa==4.9.1
s3transfer==0.10.4
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.36
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.32.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.1
